zxcdc

+++ src/Model/Table/InvoiceRowsTable.php
@@ -74,7 +74,7 @@ class InvoiceRowsTable extends Table
             ->allowEmpty('id', 'create');
 		
         $validator
-            ->integer('quantity')
+            ->decimal('quantity')
             ->requirePresence('quantity', 'create')
             ->notEmpty('quantity');
